requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
dataclass==0.0
idna==2.9
imutils==0.5.3
joblib==0.14.1
keyboard==0.13.5
kiwisolver==1.2.0
lor-deckcodes==1.0.2
matplotlib==3.2.1
MouseInfo==0.1.3
numpy==1.18.4
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
ortools==7.6.7691
Pillow==7.1.2
protobuf==3.11.3
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pywin32==227
requests==2.23.0
scapy==2.4.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tesserocr==2.4.0
typing==3.7.4.1
urllib3==1.25.9
uuid==1.30
